% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_jacc.R
\name{calc_jacc}
\alias{calc_jacc}
\title{Calculate cluster stability for each cluster}
\usage{
calc_jacc(
  fs,
  fsom_orig,
  scale,
  seed = NULL,
  boot,
  colsToUse,
  nClus,
  stim_and_prog_equal = TRUE,
  sample_by = "cell and sample"
)
}
\arguments{
\item{fs}{flowSet. Assumed appropriately transformed (but not necessary scaled).}

\item{fsom_orig}{output from FlowSOM::FlowSOM, as applied to original data.}

\item{scale}{logical. If \code{TRUE}, the variables are centred and scaled.}

\item{seed}{integer. If not \code{NULL}, then the seed is set to this.}

\item{boot}{integer. Number of bootstrap samples to take.}

\item{colsToUse, nClus}{same meaning as in FlowSOM::FlowSOM.}

\item{stim_and_prog_equal}{logical. If \code{TRUE}, then for each stimulation
and progressor combination an equal number of cells are sampled as in the original dataset.}

\item{sample_by}{'cell and sample'. If \code{'cell and sample'}, we randomly select a sample with replacement, 
and then randomly select cells from each sample with replacement, until we have
the same number of cells as in the original dataset.}
}
\value{
A tibble::tibble with a column for each cluster, where the i-th
entry for a particular column is the jaccard coefficient for that cluster for the i-th bootstrapped run.
}
\description{
Calculate cluster stability for each cluster
}
