% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_objects.R
\docType{data}
\name{dir_faust}
\alias{dir_faust}
\title{FAUST project directory}
\format{
An object of class \code{character} of length 1.
}
\usage{
dir_faust
}
\description{
FAUST project directory
}
\keyword{datasets}
