% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowsom-dim_red.r
\name{.plot_dim_red_clust_rds}
\alias{.plot_dim_red_clust_rds}
\title{Plot dim red view of data}
\usage{
.plot_dim_red_clust_rds(
  flowsom_out,
  stim_plot,
  clust,
  method,
  col_clust = NULL,
  font_size = 10,
  inc_exc_stim
)
}
\arguments{
\item{col_clust}{character.
Colour of in-cluster points.
Default is \code{"red"}.
Out-of-cluster points are always \code{"gray75"}.}

\item{inc_exc_stim}{logical.
If \code{TRUE}, then non-cluster
cells not from the stim condition of concern are plotted.
If \code{FALSE}, then they are excluded (and only cells
that belong to the stim are plotted).}
}
\description{
Plot dimension-reduced view of data,
for all stims and UMAP and t-SNE and all clusters
individually.
}
