% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowsom_results.R
\name{.plot_flowsom}
\alias{.plot_flowsom}
\title{Plot informative plots to assist cluster interpretation for an individual set of stimulations}
\usage{
.plot_flowsom(
  flowsom_out,
  flowsom_freq,
  dir_save,
  chnl_lab,
  chnl_sel,
  stim_plot,
  dim_red = TRUE
)
}
\arguments{
\item{stim_plot}{character vector. Stimulations to plot. If 'uns' and other stimulations are also in flowsom_out$stim, then it will be plotted 
where appropriate as well.}

\item{cell_count_lab}{named numeric vector. Each name indicates sample id and stimulation (paste0(SubjectID, "_", VisitType, 
"_", stim)), and each element is the number of classified cells for that combination of sample id and stimulation.}

\item{stim_clust}{character: 'all', 'all_u' or individual stim. Specifies stimulations that were used to cluster.}
}
\description{
Plot informative plots to assist cluster interpretation for an individual set of stimulations
}
