% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowsom-fs_cluster_freq.r
\name{.plot_freq}
\alias{.plot_freq}
\title{Plot cluster}
\usage{
.plot_freq(
  .data,
  x,
  y,
  dir_save,
  x_dodge = NULL,
  fill_col = NULL,
  axis_lab = NULL,
  font_size = 14,
  facet_scale = c("free", "asn", "fixed"),
  fill_lab,
  trans_y = "identity",
  n_col = NULL,
  height,
  width
)
}
\description{
Plot cluster
}
