% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_flowsom.R
\name{fit_flowsom}
\alias{fit_flowsom}
\title{Fit FlowSOM to flowSet}
\usage{
fit_flowsom(
  fs,
  n_clust,
  scale,
  chnl_sel,
  ds_name,
  dir_save_base,
  jacc_reuse = FALSE,
  jacc_n = 0,
  fs_seed = NULL,
  jacc_seed = NULL,
  chnl_lab_vec,
  tsne_tbl,
  stim
)
}
\arguments{
\item{fs}{flowSet.}

\item{n_clust}{"n_m" or "n_f". If \code{"n_m"}, then 15 clusters are used. 
If \code{"n_f"}, then 12 clusters are used for CD4 and CD8 datasets but only 7 for 
the TCRgd dataset.}

\item{scale}{'scale' or 'n_s'. If \code{scale}, then columns are centred and standardised.}

\item{chnl_sel}{character vector. Channels to apply FlowSOM to.}

\item{ds_name}{character. Name of dataset (as used to name datasets in output of cytofacs package).}

\item{dir_save_base}{character. Base directory to save to.}

\item{jacc_reuse}{logical. If \code{TRUE}, then old data is reused if available. Default is \code{FALSE}.}

\item{jacc_n}{Number of bootstrapped samples to apply cluster algorithm to to calculate cluster stability. Default is 0.}

\item{fs_seed, jacc_seed}{integer. Seeds to use for the FlowSOM and cluster stability algorithms.}

\item{chnl_lab_vec}{named character vector. If not \code{NULL}, channels are renamed to markers using it. 
Default is \code{NULL}.}

\item{tsne_tbl}{dataframe. Has columns SubjectID, VisitType, tsne1, tsne2, umap1 and umap2.}

\item{stim}{'all_u', 'all', 'p1', 'p4', 'ebv' or 'mtbaux'. Specifies stimulations.}
}
\value{
A dataframe with cells along rows and cell-level information ranging from sample information to
cluster and tsne information to expression information.
}
\description{
Also calculates cluster stability and extracts sample-specific information.
}
