% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dir_fcs.R
\name{get_dir_fcs_cyt}
\alias{get_dir_fcs_cyt}
\title{Get directory in which FCS files are stored}
\usage{
get_dir_fcs_cyt(dir_gating, pop, chnl, exc)
}
\arguments{
\item{dir_gating}{character. Directory in which
gating results are stored, i.e. the last
part should be "gating".}

\item{pop}{'cd4', 'cd8' or 'tcrgd'. Cell population.}

\item{chnl}{character vector. Channels that should be
gated together.}

\item{exc}{list of one list. Each element list should specify
the cytokines to be excluded.}
}
\description{
Get directory in which FCS files are stored
}
\examples{
# no excluded cyt combns
get_dir_fcs_cyt(
  dir_gating = dir_cytofacs_gt,
  pop = 'cd4',
  chnl = c('Ho165Di', 'Gd158Di', 'Nd146Di', 'Dy164Di'),
  exc = list(NULL))
# exclude single TNF
get_dir_fcs_cyt(
  dir_gating = dir_cytofacs,
  pop = 'cd4',
  chnl = c('Ho165Di', 'Gd158Di', 'Nd146Di', 'Dy164Di'),
  exc = list(list("Nd146Di")))
}
