% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dir_flowsom.R
\name{get_dir_flowsom_cyt}
\alias{get_dir_flowsom_cyt}
\title{Get path to FlowSOM results\\}
\usage{
get_dir_flowsom_cyt(dir_gating, pop, chnl, exc, non_responders, stim, n_clust)
}
\description{
Get the path to the directory containing 
FlowSOM results for a given dataset 
under a differernt parameterisation.
}
