% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_flowsom.R
\name{get_expr_flowsom}
\alias{get_expr_flowsom}
\alias{ex_tbl_cyt}
\title{Get expression dat with dim-red, prob and FlowSOM dat}
\usage{
get_expr_flowsom(dir_gating, pop, chnl, exc, ind = NULL)

ex_tbl_cyt(dir_gating, pop, chnl, exc, ind = NULL)
}
\arguments{
\item{dir_gating}{character. Directory in which
gating results are stored, i.e. the last
part should be "gating".}

\item{pop}{'cd4', 'cd8' or 'tcrgd'. Cell population.}

\item{chnl}{character vector. Channels that should be
gated together.}

\item{exc}{list of one list. Each element list should specify
the cytokines to be excluded.}

\item{ind}{integer vector. Indices of FCS files to load. Number of FCS files to load. If \code{NULL},
then all files are loaded. Default is \code{NULL}.}
}
\value{
A tibble with columns corresponding to marker names, with
columns specifying the FCS name and sample info.
}
\description{
Get expression dat with dim-red, prob and FlowSOM dat
}
