% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fs_cluster_post_probs.R
\name{get_fs_cluster_post_probs}
\alias{get_fs_cluster_post_probs}
\title{Get posterior probability of a response for each cluster}
\usage{
get_fs_cluster_post_probs(
  flowsom_freq = flowsom_freq,
  dir_save_cluster_results = NULL,
  reuse = FALSE
)
}
\arguments{
\item{dir_save_cluster_results}{character. Directory to save output to, if created, and to get past output from (if reuse == TRUE).}

\item{reuse}{logical. If \code{TRUE} (and dir_save != NULL), then previous output will be reused if found in 
\code{dir_save}.}
}
\value{
If there is an 'uns' stim, thena dataframe with columns clust, stim
SubjectID, VisitType, Progressor and prob is returned. 
If there is no 'uns' stim, then NULL is returned.
}
\description{
Get posterior probability of a response for each cluster
}
\details{
If the MIMOSA fit fails for a given stim and cluster combination
with the EM algorithm, then MCMC will
be attempted. If this too fails, then a single row dataframe
with NAs except for stim and cluster will be returned for that stim
and cluster combination.
}
