% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowsom_quantile_grid.R
\name{get_order_list}
\alias{get_order_list}
\title{Get list of vectors ordering markers and clusters along heat map}
\usage{
get_order_list(flowsom_out, ecdf_list = NULL, chnl_lab, chnl_sel = NULL)
}
\description{
Get list of vectors ordering markers and clusters along heat map
}
