% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fs_for_flowsom.R
\name{load_fs_for_flowsom}
\alias{load_fs_for_flowsom}
\title{Load cytokine-positive expression data for FlowSOM}
\usage{
load_fs_for_flowsom(
  dir_fcs,
  stim,
  stim_vec_u = NULL,
  responders_list,
  chnl_sel = NULL,
  uns_chr = "uns",
  conv_mtbaux_to_mtb = FALSE,
  trans = flowCore::arcsinhTransform(b = 1/5, a = 0)
)
}
\arguments{
\item{dir_fcs}{character. Directory containing FCS files.}

\item{stim}{'all_u', 'all', 'p1', 'p4', 'ebv' or 'mtbaux'. Specifies stimulation(s) that should be read in together 
as one flowSet.}

\item{stim_vec_u}{character vector. Names of all stimulations that could be read in together.}

\item{responders_list}{named list. Named list, where names are stimulations and each element is a character vector 
of samples that respond.}

\item{chnl_sel}{character vector. Channel names to transform, if transformation is supplied.}

\item{uns_chr}{character. Name for unstim in \code{stim_vec_u}. Default is "uns".}

\item{conv_mtbaux_to_mtb}{logical. If \code{TRUE}, then when matching fcs files to responder lists, 
"mtbaux"  is converted to "mtb". Default is \code{FALSE}.}

\item{trans}{object of class 'transform'. Transformation to appy to data. If \code{NULL}, then no transformation is applied.
Default is \code{flowCore::arcsinhTransform(b = 1/5,a = 0)}, 
which corresponds to asinh(x/5).}
}
\description{
Loads flowSet for flowSOM, automatically filtering out non-responders (if desired), automatically 
excluding files that didn't read in correctly and
}
\details{

}
