% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowsom-heatmap.R
\name{plot_clust_map}
\alias{plot_clust_map}
\title{Plot heat map of marker expression per cluster}
\usage{
plot_clust_map(
  flowsom_out,
  path_flowsom_out,
  dir_save,
  stim_plot,
  even_col = TRUE,
  expand_mid_col = FALSE,
  stability_show = FALSE,
  stability_min = 0,
  col_scheme = "RdYlBu",
  chnl_lab,
  chnl_sel = NULL
)
}
\arguments{
\item{stability_show}{logical.
If TRUE, then stability of each cluster is shown
beneath its text on the x-axis.}

\item{stability_min}{[0,1].
Only clusters with stability greater than this value are shown.}
}
\description{
Plot heat map of marker expression per cluster
}
