% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowsom-dim_red.r
\name{plot_dim_red}
\alias{plot_dim_red}
\title{Plot t-SNE and UMAP dim reds with clusters highlighted}
\usage{
plot_dim_red(
  flowsom_out,
  path_flowsom_out,
  dir_save,
  stim_plot,
  col_clust,
  font_size = 14,
  width = 6,
  height = 6
)
}
\arguments{
\item{col_clust}{character.
Colour of in-cluster points.}
}
\description{
Plot t-SNE and UMAP dim reds with clusters highlighted
}
