% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowsom.R
\name{plot_flowsom}
\alias{plot_flowsom}
\title{Plot FlowSOM results}
\usage{
plot_flowsom(
  flowsom_out,
  flowsom_freq,
  flowsom_cluster_post_probs,
  stim,
  dir_save_base,
  chnl_lab,
  stats_combn_tbl,
  uns_chr = "uns",
  reuse = FALSE,
  dir_save_cluster_results = NULL,
  chnl_sel
)
}
\arguments{
\item{chnl_lab}{named character vector.
If not \code{NULL}, channels are renamed to markers using it.
Default is \code{NULL}}

\item{stats_combn_tbl}{dataframe.
Contains total classified cells for each sample
(SubjectID and VisitType) for
each stimulation in a column called \code{n_cell_stim}
and also has the total number of classified cells
in the corresponding unstim sample in a column
called \code{n_cell_uns}.}

\item{uns_chr}{character.
Name for unstim in \code{stim_vec_u}.
Default is "uns".}

\item{reuse, dir_save_cluster_results}{logical and character, respectively.
If \code{reuse == TRUE} and \code{!is.null(dir_save_cluster_results)},
then if flowsom_freq is the same as flowsom_freq found}

\item{responders_list}{list.
Output from\code{get_post_probs} function for the current combination
of excluded cytokine combinations, gate method and
stimulation (all, all_u or a specific stim).}
}
\description{
Plot FlowSOM results
}
