% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowsom_quantile_grid.R
\name{plot_flowsom_grid_quantile}
\alias{plot_flowsom_grid_quantile}
\title{Plot heat map of marker expression per cluster}
\usage{
plot_flowsom_grid_quantile(
  flowsom_out,
  order_list = NULL,
  ecdf_list = NULL,
  even_col = TRUE,
  expand_mid_col = FALSE,
  stability_show = FALSE,
  stability_min = 0,
  col_scheme = "RdYlBu",
  chnl_lab,
  chnl_sel = NULL,
  stim_plot,
  dir_save
)
}
\arguments{
\item{stability_show}{logical. If TRUE, then stability of each cluster is shown 
beneath its text on the x-axis.}

\item{stability_min}{[0,1]. Only clusters with stability greater than this value are shown.}
}
\description{
Plot heat map of marker expression per cluster
}
