% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowsom-fs_cluster_freq.r
\name{plot_freq}
\alias{plot_freq}
\title{Plot cluster frequencies per cyt+ and all cells, by stim and prog}
\usage{
plot_freq(data, data_no_uns, dir_save, fill_col, fill_lab)
}
\arguments{
\item{data, data_no_uns.}{Whatever works with \code{.data} parameter of \code{saver::save_rds_eval}.}

\item{dir_save}{character. Directory to save to.}
}
\description{
Plot cluster frequencies per cyt+ and all cells, by stim and prog
}
