% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowsom_results.R
\name{plot_fs_cluster_freq}
\alias{plot_fs_cluster_freq}
\title{Plot cluster frequencies per cyt+ and all cells, by stim and prog}
\usage{
plot_fs_cluster_freq(flowsom_freq, dir_save)
}
\description{
Plot cluster frequencies per cyt+ and all cells, by stim and prog
}
