% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowsom-dens.R
\name{plot_marker_density_by_cluster}
\alias{plot_marker_density_by_cluster}
\title{Plot marker density for each cluster}
\usage{
plot_marker_density_by_cluster(
  flowsom_out,
  path_flowsom_out,
  chnl_lab,
  dir_save,
  stim_plot,
  font_size
)
}
\arguments{
\item{stim_plot}{character: "all", "all_u", or individual stim.
If "all" or "all_u", then
all cells in flowsom_out belonging to a particular cluster
are plotted for that cluster, including unstim.
If an individual stim, then only cells belonging
to that stim are plotted for that cluster.
When plotting non-cluster cells, all
non-cluster cells are plotted regardless of the value for
\code{stim_clust}.}
}
\description{
Plot marker density for each cluster
}
