% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cytokine_data.R
\name{preprocess_cytokine_data}
\alias{preprocess_cytokine_data}
\title{Preprocess cytokine data}
\usage{
preprocess_cytokine_data(
  path_faust_dir,
  path_output,
  gs_nm = "gs_cytof_acs_nk",
  chnls_gate = c("Ho165Di", "Gd158Di", "Nd146Di", "Dy164Di", "Gd156Di", "Nd150Di"),
  chnls_analyse = "",
  gate = FALSE,
  fast = FALSE,
  stats_only = FALSE,
  plots_only = FALSE,
  save_cyt_pos_fcs = TRUE,
  combn_exc_list = list(NULL, c("Nd146Di", "Gd156Di", "Nd150Di")),
  hladr = FALSE,
  hladr_bulk = FALSE,
  label_cytokines = PipelineDataACSCyTOF::label_cytokines,
  compass = TRUE,
  compass_stim = c("p1", "mtb", "ebv", "p4"),
  compass_fast = FALSE,
  downstream_exc_list = NULL,
  tsne = TRUE,
  tsne_plot = TRUE,
  tsne_plot_force = TRUE,
  tsne_save = TRUE,
  tsne_reuse = TRUE,
  post_probs_bulk = TRUE,
  post_probs_bulk_stim = c("p1", "mtb", "ebv"),
  flowsom = TRUE,
  flowsom_stim = c("all_u"),
  flowsom_p4_exc = TRUE,
  flowsom_scale = FALSE,
  flowsom_n_clust = "n_f",
  flowsom_r_o = c("r_o", "all"),
  jacc_n = 100,
  jacc_reuse = TRUE,
  flowsom_post_probs = FALSE,
  flowsom_plot = TRUE,
  flowsom_plot_force = TRUE
)
}
\arguments{
\item{path_faust_dir}{character. Path to FAUST project directory.}

\item{label_cytokines}{function. A function that takes
the cytokine names as input and returns the
value desired to be plotted along the y-axis
of the grid plot.
If not \code{NULL}, then
it is supplied when creating the grid plot to
the \code{scale_y_continuous} function
via the \code{labels} parameter.
For example, if we have two cytokines,
\code{"IFNg"} and
\code{"TNF"}, but we want to
display \code{"IFNg"} with
the Greek gamma symbol, then we can
set \code{cyt_lab} equal to the following:
\code{cyt_lab = function(cyt)
purrr::map(cyt, function(cyt_ind) {switch(cyt_ind,
"IFNg" = bquote(paste(plain(paste("IFN")), gamma)),
cyt_ind)})}.
This will change the label for
\code{"IFNg"} but leave all the others as is.}
}
\description{
Preprocess cytokine data
}
