% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_list.r
\name{save_list}
\alias{save_list}
\title{Save objects in a list}
\usage{
save_list(..., obj_list, ftype, height = 6, width = 6, scale = 1, dir_save)
}
\arguments{
\item{...}{name-value pairs. Saved.}

\item{obj_list}{named list. If not named, then names
will be provided.}

\item{ftype}{"pdf", "png" and/or "rds". Save format.}

\item{height, width, scale}{numeric. Passed to \code{cowplot::ggsave}.}

\item{dir_save}{directory. To save files to.}
}
\value{
\code{invisible(TRUE)}.
}
\description{
Save multiple objects of multiple classes
in multiple formats with less error-prone naming
method of using a named list.
}
