% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowsom_results.R
\name{plot_dim_red}
\alias{plot_dim_red}
\title{Plot t-SNE and UMAP dim reds with clusters highlighted}
\usage{
plot_dim_red(flowsom_out, dir_save, stim_plot)
}
\description{
Plot t-SNE and UMAP dim reds with clusters highlighted
}
